+++ src/pages/index.tsx
@@ -27,7 +27,7 @@ const Home: NextPage = () => {
           onClick={(newStar) => setStars(newStar)}
         /> */}
         {posts?.length && (
-          <div className='mt-4 grid gap-6 grid-cols-[repeat(auto-fill,minmax(291px,auto))] justify-center'>
+          <div className='mt-4 grid gap-6 grid-cols-[repeat(auto-fill,minmax(291px,auto))] justify-center items-start'>
             {posts.map((post, i) => (
               <PostItem post={post} key={i} className='mb-1' />
             ))}
